CORE-1434: HostController -> NodeApiController
